lfht: use a maximum probe depth
Applied in ht_add() and ht_val(), as it should be. This makes t/21_* run in
under 100 ms on a 8-way Piledriver, which is rather a drastic improvement over
previous performance.

On the downside, this adds an abort() in a path that can neither be excluded
by analysis or hit in a unit test (so far).