t: add 22_lfht_multi_mt
This tests concurrent lfht operation as a multiset.

In particular, the test tries to detect the case where an entry is deleted
from underneath a migrator's feet, and the migrated entry is also deleted
before the migrator can undo it. This should cause either failing deletions,
or entries that stick around despite the number of completed adds being equal
to the number of succeeding deletes.

For now, it hits some bugs sometimes. Another test methodology is required to
hit particular cases reliably.