lfht: add robust migration protocol ``SMHED''
The algorithm reserves five more (titular) bits from common_mask and uses them
to replace the constant tombstone values, and to implement a migration and
deletion protocol. It allows concurrent migrations, deletions, and lookups
while ensuring

  1) visibility of items that should be present;
  2) concurrent deletion of the same item from multiple threads succeeding
     exactly once;
  3) number of succeeding deletions on a key being at most the number of
     previously completed adds;
  4) that once all copies of a key have been deleted, they key will become
     undiscoverable in the hash table once all concurrent migrations on it
     have completed;
  5) the migration algorithm's robustness wrt ABBA races.

In particular this changeset fixes the problem where concurrent deletions and
migrations would sometimes undo deletions or cause them to succeed where they
shouldn't have. There was an additional ABBA race where a slot would be
deleted and then recycled to the same value in the middle of a migration; this
is also corrected.

README.md was updated somewhat, with proper documentation to follow. Various
cleanups too.

(for example, currently resv_mask is the wrong name and includes perfect_bit,
which makes things more difficult than they need to be. and the slot
predicates are messed up.)