lfht: dereference migration pointers in lfht_delval()
The alive ones are dereferenced straight up as an optimization (the gen_id
part allows omission of multiple table probes), and a new candidate entry
sought if the pointer is already empty.

Idea here is that if clients arbitrate for consumption of a data structure,
then this should also imply a successful lfht_delval() of that address for the
sole consumer. This is necessary because entries may be returned from
iteration for a while after their final lfht_delval() success has been
reported.