lfht: broadcast detected conditions in ht_full_test()
Before this, other threads would keep filling the table up until each had
observed a per-cpu interval. That was wrong; the table should be left alone
as soon as possible. To that end we'll reset the intervals for other CPUs and
not increment ours; they'll do the same.

As a result, all unit tests now run as quickly as one would expect of a hash
table. There's still variance but its magnitude is far less than multiple
seconds.